+++ docs/src/examples/components/Menu/index.tsx
@@ -2,11 +2,13 @@ import * as React from 'react'
 import Types from './Types'
 import States from './States'
 import Variations from './Variations'
+import Content from './Content'
 
 const MenuExamples = () => (
   <div>
     <Types />
     <States />
+    <Content />
     <Variations />
   </div>
 )