+++ src/themes/teams/components/Menu/menuItemStyles.ts
@@ -1,4 +1,5 @@
-import { pxToRem, getSideArrow } from '../../utils'
+import { getSideArrow } from '../../utils'
+import { pxToRem } from '../../../../lib'
 import { ComponentSlotStyleFunction, ComponentSlotStylesInput, ICSSInJSStyle } from '../../../types'
 import { MenuVariables } from './menuVariables'
 import { MenuItemProps, MenuItemState } from '../../../../components/Menu/MenuItem'
@@ -141,7 +142,7 @@ const pointingBeak: ComponentSlotStyleFunction<MenuItemPropsAndState, MenuVariab
 }
 
 const menuItemStyles: ComponentSlotStylesInput<MenuItemPropsAndState, MenuVariables> = {
-  wrapper: ({ props, variables: v, theme }): ICSSInJSStyle => {
+  wrapper: ({ props, variables: v, theme, colors }): ICSSInJSStyle => {
     const {
       active,
       disabled,
@@ -198,7 +199,7 @@ const menuItemStyles: ComponentSlotStylesInput<MenuItemPropsAndState, MenuVariab
         display: 'flex',
       }),
 
-      ...itemSeparator({ props, variables: v, theme }),
+      ...itemSeparator({ props, variables: v, theme, colors }),
 
       // active styles
       ...(active && {
@@ -209,7 +210,7 @@ const menuItemStyles: ComponentSlotStylesInput<MenuItemPropsAndState, MenuVariab
             ? pointing === 'end'
               ? { borderRight: `${pxToRem(3)} solid ${v.primaryActiveBorderColor}` }
               : { borderLeft: `${pxToRem(3)} solid ${v.primaryActiveBorderColor}` }
-            : pointingBeak({ props, variables: v, theme }))),
+            : pointingBeak({ props, variables: v, theme, colors }))),
       }),
 
       // focus styles